refactor: move audit log packets to sub namespace
